feat(templates): split web_dev.py
