+++ autograder/template_library/__init__.py
@@ -1,6 +1,6 @@
 """Template library package exports."""
 
-from autograder.template_library.web_dev import WebDevTemplate
+from autograder.template_library.web_dev.template import WebDevTemplate
 from autograder.template_library.api_testing import ApiTestingTemplate
 from autograder.template_library.input_output import InputOutputTemplate
