+++ autograder/template_library/web_dev/__init__.py
@@ -1,0 +1,1 @@
+#
